feat: Controllers are ranged by domain and their models are included side-by-side
