+++ src/apps/server/Afina2.Server.WebApp/Controllers/Sysinfos/Endpoint.cs
@@ -1,4 +1,5 @@
-public class EndpointVM
+namespace Afina2.Server.WebApp.Controllers.Sysinfos;
+public class Endpoint
 {
     public string? Method { get; set; }
     public string? Route { get; set; }
