+++ src/apps/server/Afina2.Server.WebApp/README.md
@@ -1,0 +1,3 @@
+# Afina WebApp
+
+API Web app for Afina platform.
